pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"
django = "^5.1.4"
pylint-django = "^2.6.1"

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.3"
pytest = "^8.3.4"
mypy = "^1.14.1"
isort = "^5.13.2"
black = "^24.10.0"
drf-spectacular = "^0.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
extend-exclude = "migrations/"

[tool.isort]
profile = "black"
extend_skip_glob = "migrations"
line_length = 79
lines_after_imports = 2

[tool.pylint]
recursive="y"
ignore = ["migrations", "__pycache__", "manage.py", "settings.py", ".venv", "asgi.py"]
load-plugins=["pylint_django"]
django-settings-module=["app"]
max-line-length = 79
disable = ["C0114", "C0115", "C0116", "W0511", "R0901"]


# mypy
[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = ["backend_clean_environment.settings"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE="backend_clean_environment.settings"
# python_files = ["test_*.py"]
# python_functions = ["test_*"]
# testpaths = ["tests"]
addopts = "-v -x"